Plans 1-4: ML rebuild, historical data, backtest consolidation, paper trading wire-up

+++ QuantPlatformApplication/src/main/java/com/QuantPlatformApplication/QuantPlatformApplication/model/dto/MarketDataResponse.java
@@ -28,5 +28,5 @@ public class MarketDataResponse {
     private BigDecimal high;
     private BigDecimal low;
     private BigDecimal close;
-    private Long volume;
+    private BigDecimal volume;
 }
